Fix user identity issue with UserFromId mode
