+++ src/TokenBinding/IEasyAuthClient.cs
@@ -5,12 +5,15 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
 namespace Microsoft.Azure.WebJobs.Extensions.AuthTokens
 {
+    using System.IdentityModel.Tokens.Jwt;
     using System.Threading.Tasks;
 
     internal interface IEasyAuthClient
     {
-        Task<EasyAuthTokenStoreEntry> GetTokenStoreEntry(TokenAttribute attribute);
+        Task<EasyAuthTokenStoreEntry> GetTokenStoreEntry(JwtSecurityToken jwt, TokenAttribute attribute);
 
-        Task RefreshToken(TokenAttribute attribute);
+        Task RefreshToken(JwtSecurityToken jwt, TokenAttribute attribute);
+
+        string GetBaseUrl();
     }
 }
