+++ test/TokenExtension.Tests/TokenTests.cs
@@ -39,6 +39,10 @@ namespace Microsoft.Azure.WebJobs.Extensions.Token.Tests
             var config = new AuthTokenExtensionConfig();
             var mockClient = GetEasyAuthClientMock(currentToken);
             config.EasyAuthClient = mockClient.Object;
+            config.AppSettings = GetNameResolver(new Dictionary<string, string>()
+            {
+                { Constants.AppSettingWebsiteAuthSigningKey, SigningKey }
+            }).Object;
 
             var args = new Dictionary<string, object>
             {
@@ -61,6 +65,10 @@ namespace Microsoft.Azure.WebJobs.Extensions.Token.Tests
             var config = new AuthTokenExtensionConfig();
             var mockClient = GetEasyAuthClientMock(expiredToken, refreshedToken);
             config.EasyAuthClient = mockClient.Object;
+            config.AppSettings = GetNameResolver(new Dictionary<string, string>()
+            {
+                { Constants.AppSettingWebsiteAuthSigningKey, SigningKey }
+            }).Object;
 
             var args = new Dictionary<string, object>
             {
@@ -71,7 +79,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.Token.Tests
 
             var expectedResult = refreshedToken.AccessToken;
             Assert.Equal(expectedResult, finalTokenValue);
-            mockClient.Verify(client => client.RefreshToken(It.IsAny<TokenAttribute>()), Times.AtLeastOnce());
+            mockClient.Verify(client => client.RefreshToken(It.IsAny<JwtSecurityToken>(), It.IsAny<TokenAttribute>()), Times.AtLeastOnce());
             ResetState();
         }
 
@@ -131,7 +139,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.Token.Tests
                 Audience = "https://sample.com",
                 Issuer = "https://sample.com",
                 Subject = identity,
-                SigningCredentials = new EasyAuthTokenClient.HmacSigningCredentials(SigningKey),
+                SigningCredentials = new EasyAuthTokenManager.HmacSigningCredentials(SigningKey),
             };
             string accessToken = jwtHandler.CreateJwtSecurityToken(descr).RawEncryptedKey;
             return new EasyAuthTokenStoreEntry()
@@ -145,7 +153,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.Token.Tests
         {
             var clientMock = new Mock<IEasyAuthClient>();
             var responseQueue = new Queue<EasyAuthTokenStoreEntry>(responsesInOrder);
-            clientMock.Setup(client => client.GetTokenStoreEntry(It.IsAny<TokenAttribute>()))
+            clientMock.Setup(client => client.GetTokenStoreEntry(It.IsAny<JwtSecurityToken>(), It.IsAny<TokenAttribute>()))
                 .Returns(Task.FromResult(responseQueue.Dequeue()));
             return clientMock;
         }
@@ -158,6 +166,16 @@ namespace Microsoft.Azure.WebJobs.Extensions.Token.Tests
             clientMock.Setup(client => client.GetTokenOnBehalfOfUserAsync(It.IsAny<string>(), It.IsAny<string>()))
                 .Returns(Task.FromResult(AccessTokenFromUserToken));
             return clientMock;
+        }
+
+        private static Mock<INameResolver> GetNameResolver(Dictionary<string, string> appSettings)
+        {
+            var mock = new Mock<INameResolver>();
+            foreach(var appSetting in appSettings)
+            {
+                mock.Setup(resolver => resolver.Resolve(appSetting.Key)).Returns(appSetting.Value);
+            }
+            return mock;
         }
 
         private class TokenFunctions
